# Speed up Write-Host