Write-Host is creating an InformationRecord.

As part of the initialization of InformationRecord, we were initializing the ComputerName property.

This is apparently expensive, so I deferred initializing the value until the first time the property is accessed, which is uncommon, but will happen during remoting (because all properties are serialized).

In my local runs, CI tests ran much faster, 254s with the change, 305s without. A similar improvement is seen in calling Write-Host repeatedly - 2.2s for 4000 calls with the change, 47s without.